Add API endpoint for listing BlogCategory

+++ src/api/paging.rs
@@ -5,8 +5,8 @@ use crate::consts::DEFAULT_PAGE_SIZE;
 use crate::utils::urls::update_entry_in_query;
 
 pub fn gen_pagination_links(paging: &Paging, total: usize, original_uri: Uri) -> PaginationLinks {
-    let current_page = paging.page.unwrap_or(1);
-    let per_page = paging.per_page.unwrap_or(DEFAULT_PAGE_SIZE);
+    let current_page = paging.page.unwrap_or(1) as usize;
+    let per_page = paging.per_page.unwrap_or(DEFAULT_PAGE_SIZE) as usize;
     let prev_page = if current_page > 1 {
         Some(current_page - 1)
     } else {
