+++ src/consts.rs
@@ -1,2 +1,2 @@
 pub const DB_NAME: &str = "quanweb";
-pub const DEFAULT_PAGE_SIZE: usize = 10;
+pub const DEFAULT_PAGE_SIZE: u8 = 10;
